quiet annoying GCC warnings
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@29328 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ samples/text/text.cpp
@@ -772,7 +772,7 @@ void MyTextCtrl::OnKeyDown(wxKeyEvent& event)
                         line, column,
                         (long) GetNumberOfLines(),
                         (long) GetLineLength(line),
-                        GetValue().length(),
+                        (unsigned int) GetValue().length(),
                         GetLastPosition());
 
                 long from, to;
@@ -782,7 +782,8 @@ void MyTextCtrl::OnKeyDown(wxKeyEvent& event)
 
                 wxLogMessage(_T("Selection: from %ld to %ld."), from, to);
                 wxLogMessage(_T("Selection = '%s' (len = %u)"),
-                             sel.c_str(), sel.length());
+                             sel.c_str(), 
+							 (unsigned int) sel.length());
             }
             break;
